refactor(component): refactor component folder structure
