+++ packages/web/src/common/components/Footer/_atomic/SPARCSLogo.tsx
